refactor: move component files to separate folders
