+++ src/components/app/app.js
@@ -1,0 +1,27 @@
+import AppHeader from "../app-header/app-header";
+import SearchPanel from "../search-panel/search-panel";
+import TodoList from "../todo-list/todo-list";
+import ItemStatusFilter from "../item-status-filter/item-status-filter";
+
+import "./app.css";
+
+const App = () => {
+  const todoData = [
+    { label: "Drink Coffee", important: false, id: 1 },
+    { label: "Make Awesome App", important: true, id: 2 },
+    { label: "Have a lunch", important: false, id: 3 },
+  ];
+
+  return (
+    <div className="todo-app">
+      <AppHeader toDo={1} done={2} />
+      <div className="top-panel d-flex">
+        <SearchPanel />
+        <ItemStatusFilter />
+      </div>
+      <TodoList todos={todoData} />
+    </div>
+  );
+};
+
+export default App;
