+++ src/components/todo-list/todo-list.js
@@ -1,4 +1,4 @@
-import TodoListItem from "./todo-list-item";
+import TodoListItem from "../todo-list-item/todo-list-item";
 import "./todo-list.css";
 
 const TodoList = ({ todos }) => {
