+++ src/index.js
@@ -1,29 +1,5 @@
 import ReactDOM from "react-dom";
 
-import AppHeader from "./components/app-header";
-import SearchPanel from "./components/search-panel";
-import TodoList from "./components/todo-list";
-import ItemStatusFilter from "./components/item-status-filter";
-
-import "./index.css";
-
-const App = () => {
-  const todoData = [
-    { label: "Drink Coffee", important: false, id: 1 },
-    { label: "Make Awesome App", important: true, id: 2 },
-    { label: "Have a lunch", important: false, id: 3 },
-  ];
-
-  return (
-    <div className="todo-app">
-      <AppHeader toDo={1} done={2} />
-      <div className="top-panel d-flex">
-        <SearchPanel />
-        <ItemStatusFilter />
-      </div>
-      <TodoList todos={todoData} />
-    </div>
-  );
-};
+import App from "./components/app/app";
 
 ReactDOM.render(<App />, document.getElementById("root"));
